+++ src/GlobalPayments.Api/Builders/BaseBuilder.cs
@@ -1,4 +1,6 @@
-﻿namespace GlobalPayments.Api.Builders {
+﻿using GlobalPayments.Api.Entities;
+
+namespace GlobalPayments.Api.Builders {
     public abstract class BaseBuilder<TResult> {
         internal Validations Validations { get; set; }
 
@@ -8,6 +10,12 @@
         }
 
         public virtual TResult Execute(string configName = "default") {
+            Validations.Validate(this);
+            return default(TResult);
+        }
+
+        public virtual TResult Execute(Secure3dVersion version, string configName = "default")
+        {
             Validations.Validate(this);
             return default(TResult);
         }