+++ src/GlobalPayments.Api/Gateways/ISecure3dProvider.cs
@@ -5,6 +5,6 @@ namespace GlobalPayments.Api.Gateways {
     public interface ISecure3dProvider {
         Secure3dVersion Version { get; }
 
-        Transaction ProcessSecure3d(Secure3dBuilder builder);
+        Transaction ProcessSecure3d(Secure3dBuilder builder);        
     }
 }