+++ src/GlobalPayments.Api/PaymentMethods/PaymentInterfaces/ICardData.cs
@@ -4,6 +4,7 @@ namespace GlobalPayments.Api.PaymentMethods {
     public interface ICardData {
         bool CardPresent { get; set; }
         string CardType { get; set; }
+        string CardHolderName { get; set; }
         string Cvn { get; set; }
         CvnPresenceIndicator CvnPresenceIndicator { get; set; }
         string Number { get; set; }