+++ src/GlobalPayments.Api/ServiceConfigs/Gateways/GpApiConfig.cs
@@ -101,6 +101,8 @@ namespace GlobalPayments.Api {
 
             services.ReportingService = gateway;
 
+            services.FraudService = gateway;
+
             services.SetPayFacProvider(gateway);
             services.SetSecure3dProvider(Secure3dVersion.One, gateway);
             services.SetSecure3dProvider(Secure3dVersion.Two, gateway);