+++ src/GlobalPayments.Api/ServicesContainer.cs
@@ -3,6 +3,7 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using GlobalPayments.Api.Entities;
 using GlobalPayments.Api.Gateways;
+using GlobalPayments.Api.Services;
 using GlobalPayments.Api.Terminals;
 
 namespace GlobalPayments.Api {
@@ -39,6 +40,8 @@ namespace GlobalPayments.Api {
         internal TableServiceConnector TableServiceConnector { get; set; }
 
         internal PayrollConnector PayrollConnector { get; set; }
+
+        internal IFraudCheckService FraudService { get; set; }
 
         internal ISecure3dProvider GetSecure3DProvider(Secure3dVersion version) {
             if (_secure3dProviders.ContainsKey(version)) {
@@ -235,6 +238,14 @@ namespace GlobalPayments.Api {
             throw new ApiException("The specified configuration has not been configured for gateway processing.");
         }
 
+        internal IFraudCheckService GetFraudCheckClient(string configName) {
+            if (_configurations.ContainsKey(configName))
+            {
+                return _configurations[configName].FraudService;
+            }
+
+            throw new ApiException("The specified configuration has not been configured for fraud check.");
+        }
         internal void removeConfiguration(String configName) {
             if(_configurations.ContainsKey(configName)) {
                 ConfiguredServices config = new ConfiguredServices();